Flash: Player size fix and unsupported browser detection.
Despite the lack of a hardcoded size, the Ruffle player doesn't seem
to use the native resolution of the movie it loads, for some reason.

The player will now always attempt to use the movie's native
resolution. If the screen is smaller, it will scale while preserving
the aspect ratio. If the screen is larger, it will scale if the 'Fit'
option is enabled, otherwise it will keep the movie's resolution.

Before running the main script, the page now checks if the user's
browser supports ES6. The Ruffle script requires it to function. If
not, the browser won't be able to display the website.
You won't be forgotten, Internet Explorer.